fastapi-backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0
sqlalchemy==2.0.23
databases[postgresql]==0.8.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8

# ML/AI Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
joblib==1.3.2
pillow==10.1.0

# HTTP & API
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Additional utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Image processing for disease detection
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1

# AI & external APIs
openweathermap==1.0.1
